refactor: move `ListsPage` and `ListPage`
