+++ lib/router/router.dart
@@ -33,8 +33,8 @@ import '../view/page/gallery/gallery_edit_page.dart';
 import '../view/page/gallery/gallery_page.dart';
 import '../view/page/gallery/gallery_post_page.dart';
 import '../view/page/image_page.dart';
-import '../view/page/list_page.dart';
-import '../view/page/lists_page.dart';
+import '../view/page/list/list_page.dart';
+import '../view/page/list/lists_page.dart';
 import '../view/page/login_page.dart';
 import '../view/page/misskey_games_page.dart';
 import '../view/page/note_page.dart';
