+++ lib/view/page/list/list_page.dart
@@ -5,19 +5,19 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:misskey_dart/misskey_dart.dart';
 import 'package:share_plus/share_plus.dart';
 
-import '../../i18n/strings.g.dart';
-import '../../model/account.dart';
-import '../../model/list_settings.dart';
-import '../../model/tab_settings.dart';
-import '../../provider/api/lists_notifier_provider.dart';
-import '../../provider/api/timeline_notes_notifier_provider.dart';
-import '../../util/copy_text.dart';
-import '../../util/future_with_dialog.dart';
-import '../../util/launch_url.dart';
-import '../dialog/confirmation_dialog.dart';
-import '../dialog/list_settings_dialog.dart';
-import '../widget/note_widget.dart';
-import '../widget/paginated_list_view.dart';
+import '../../../i18n/strings.g.dart';
+import '../../../model/account.dart';
+import '../../../model/list_settings.dart';
+import '../../../model/tab_settings.dart';
+import '../../../provider/api/lists_notifier_provider.dart';
+import '../../../provider/api/timeline_notes_notifier_provider.dart';
+import '../../../util/copy_text.dart';
+import '../../../util/future_with_dialog.dart';
+import '../../../util/launch_url.dart';
+import '../../dialog/confirmation_dialog.dart';
+import '../../dialog/list_settings_dialog.dart';
+import '../../widget/note_widget.dart';
+import '../../widget/paginated_list_view.dart';
 
 class ListPage extends HookConsumerWidget {
   const ListPage({
