+++ lib/view/page/list/lists_page.dart
@@ -4,13 +4,13 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
 import 'package:misskey_dart/misskey_dart.dart';
 
-import '../../i18n/strings.g.dart';
-import '../../model/account.dart';
-import '../../provider/api/i_notifier_provider.dart';
-import '../../provider/api/lists_notifier_provider.dart';
-import '../../util/future_with_dialog.dart';
-import '../dialog/text_field_dialog.dart';
-import '../widget/error_message.dart';
+import '../../../i18n/strings.g.dart';
+import '../../../model/account.dart';
+import '../../../provider/api/i_notifier_provider.dart';
+import '../../../provider/api/lists_notifier_provider.dart';
+import '../../../util/future_with_dialog.dart';
+import '../../dialog/text_field_dialog.dart';
+import '../../widget/error_message.dart';
 
 class ListsPage extends ConsumerWidget {
   const ListsPage({super.key, required this.account});
